collaborative/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
python-dotenv
joblib
lightgbm
pandas
numpy
scikit-learn
sentence-transformers
httpx
python-multipart
aiohttp
deep-translator

